refactor: slightly better dynamic get method (#187)
