+++ include/cocaine/logging.hpp
@@ -95,4 +95,18 @@ log(Log&& log, cocaine::logging::priorities severity, Message&& message, const A
 
 }}}  // namespace cocaine::detail::logging
 
+namespace cocaine { namespace logging {
+
+// C++ typename demangling
+auto
+demangle(const std::string& mangled) -> std::string;
+
+template<class T>
+auto
+demangle() -> std::string {
+    return demangle(typeid(T).name());
+}
+
+}}  // namespace cocaine::logging
+
 #endif
