+++ src/logging.cpp
@@ -18,7 +18,7 @@
     along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "cocaine/detail/logging.hpp"
+#include "cocaine/logging.hpp"
 
 #include <cxxabi.h>
 
@@ -30,7 +30,7 @@
 
 #include "cocaine/format.hpp"
 
-namespace cocaine { namespace detail { namespace logging {
+namespace cocaine { namespace logging {
 
 std::string
 demangle(const std::string& mangled) {
@@ -58,4 +58,4 @@ demangle(const std::string& mangled) {
     return buffer.get();
 }
 
-}}} // namespace cocaine::detail::logging
+}} // namespace cocaine::logging
